Add the Next.js frontend with a 3D forest hero, and fix bugs found seeding and walking the full stack end-to-end
Frontend (Next.js 16 App Router, React 19, TypeScript strict, Tailwind v4):
- 19 routes covering every role's workflow: public landing/auth, dashboard,
  observation capture/list/detail with expert verification, species
  catalogue, GIS map, analytics, verification queue, alerts, sensor
  network, research experiments, and user administration.
- Three.js/react-three-fiber 3D scenes built specifically for this domain,
  not generic decor: an instanced canopy field (420 trees, one draw call)
  with a hand-written GLSL wind-sway vertex shader driven by per-instance
  phase and height, drifting ground mist and fireflies with their own
  noise/glow fragment shaders, and a data-driven "species orbit" widget on
  the dashboard where each node's size and orbital speed is computed live
  from the deployment's own species_by_category counts.
- Strict separation between AI-assisted identification and expert-verified
  taxonomy is carried into the UI: every AIPrediction is rendered with its
  modality, model version, and confidence, and is never treated as the
  observation's species until an EXPERT/ADMIN explicitly confirms/corrects
  it through VerifyPanel.
- Deterministic mulberry32 PRNG for all procedural 3D placement (trees,
  fireflies) instead of Math.random(), so SSR output is stable and the
  scenes satisfy the React Compiler's purity lint rule honestly rather
  than by disabling it.
- eslint-config-next's native flat-config export is used directly; the
  legacy FlatCompat/@eslint/eslintrc bridge is not used because it throws
  a circular-JSON error under current ESLint when validating
  next/core-web-vitals.

Bugs found and fixed while seeding real data and clicking through every
page against a live backend:
- pydantic-settings v2 was JSON-decoding CORS_ORIGINS/ALLOWED_*_EXTENSIONS
  before the custom comma-separated field_validator ran, crashing
  Settings() on start. Fixed by marking those fields NoDecode.
- scripts/seed.py called the AI engine's render_image() with a stdlib
  random.Random instead of the numpy.random.Generator it expects, and its
  "AI got it right" verification branch could pick a corrected_species_id
  equal to the already-reported species — which the API correctly rejects
  (that's a CONFIRM, not a CORRECT). Both fixed; a 90-observation reseed
  now produces zero warnings across all 36 verification submissions.
- SpeciesOrbit's 3D canvas rendered as an opaque white box on the
  dashboard: `<color attach="background" args={["transparent"]}>` is not a
  valid THREE.Color value. Fixed by giving the canvas a real transparent
  clear color (`gl={{alpha:true}}` + `gl.setClearColor(0x000000, 0)`) so
  the orbit blends into the dark dashboard card as intended.
- Two lint regressions surfaced by a full repo-wide ruff pass: an
  int()-around-round() no-op flagged by RUF046, and app/models/enums.py's
  hand-rolled `class StrEnum(str, Enum)` flagged by UP042 now that the
  target Python version ships enum.StrEnum natively; both fixed.

Verified before this commit: 202/202 backend+AI tests pass, `ruff check .`
is clean across backend/ai/scripts, `next lint` and `next build` are both
clean (19 routes), and every page was walked with Playwright against a
freshly seeded database (90 observations across 6 zones, all 32 species,
all 6 roles) — dashboard, observations, species, map, analytics,
verification, alerts, sensors, and admin/users all render real data
correctly. The map's OpenStreetMap tile imagery doesn't load in this
sandbox specifically because the outbound proxy's allowlist rejects
tile.openstreetmap.org (confirmed via the proxy's own status endpoint,
which logs repeated 403s to that host) — Leaflet, the marker layer, and
the privacy-generalised-coordinates notice all render correctly, and tiles
will load normally on any deployment with open internet access.

+++ ai/preprocessing/audio.py
@@ -207,7 +207,7 @@ def resample_linear(samples: np.ndarray, source_rate: int, target_rate: int) -> 
     if source_rate == target_rate or samples.size == 0:
         return samples
     duration = samples.size / float(source_rate)
-    target_length = max(1, int(round(duration * target_rate)))
+    target_length = max(1, round(duration * target_rate))
     source_positions = np.linspace(0.0, duration, num=samples.size, endpoint=False)
     target_positions = np.linspace(0.0, duration, num=target_length, endpoint=False)
     return np.interp(target_positions, source_positions, samples).astype(np.float32)
